# EVP_CIPHER_CTX_get_key_length and EVP_CIPHER_CTX_get_iv_length speedup