--- a/providers/implementations/ciphers/ciphercommon_gcm.c
+++ b/providers/implementations/ciphers/ciphercommon_gcm.c
@@ -137,16 +137,25 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
     PROV_GCM_CTX *ctx = (PROV_GCM_CTX *)vctx;
     OSSL_PARAM *p;
     size_t sz;
+    int count = OSSL_PARAM_count(params);
 
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_IVLEN);
-    if (p != NULL && !OSSL_PARAM_set_size_t(p, ctx->ivlen)) {
-        ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
-        return 0;
+    if (p != NULL) {
+        if (!OSSL_PARAM_set_size_t(p, ctx->ivlen)) {
+            ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
+            return 0;
+        }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_KEYLEN);
-    if (p != NULL && !OSSL_PARAM_set_size_t(p, ctx->keylen)) {
-        ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
-        return 0;
+    if (p != NULL) {
+        if (!OSSL_PARAM_set_size_t(p, ctx->keylen)) {
+            ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
+            return 0;
+        }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_AEAD_TAGLEN);
     if (p != NULL) {
@@ -157,6 +166,8 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
             ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
             return 0;
         }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
 
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_IV);
@@ -172,6 +183,8 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
             ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
             return 0;
         }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
 
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_UPDATED_IV);
@@ -187,12 +200,18 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
             ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
             return 0;
         }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
 
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_AEAD_TLS1_AAD_PAD);
-    if (p != NULL && !OSSL_PARAM_set_size_t(p, ctx->tls_aad_pad_sz)) {
-        ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
-        return 0;
+    if (p != NULL) {
+        if (!OSSL_PARAM_set_size_t(p, ctx->tls_aad_pad_sz)) {
+            ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
+            return 0;
+        }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_AEAD_TAG);
     if (p != NULL) {
@@ -208,6 +227,8 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
             ERR_raise(ERR_LIB_PROV, PROV_R_FAILED_TO_SET_PARAMETER);
             return 0;
         }
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
     p = OSSL_PARAM_locate(params, OSSL_CIPHER_PARAM_AEAD_TLS1_GET_IV_GEN);
     if (p != NULL) {
@@ -215,6 +236,8 @@ int ossl_gcm_get_ctx_params(void *vctx, OSSL_PARAM params[])
             || p->data_type != OSSL_PARAM_OCTET_STRING
             || !getivgen(ctx, p->data, p->data_size))
             return 0;
+        if (--count == 0)
+            return 1; /* short-cut for speedup */
     }
     return 1;
 }