This is just a draft PR for the approach I've suggested. IMO this does not really complicate the get_ctx_param code however of course it is not really generic approach and speeds up in practice only the case where we get a single param and one that is near the beginning of the params we check in the get_ctx_param code.

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->
- [ ] documentation is added or updated
- [ ] tests are added or updated