@t8m 2021-12-15
Some numbers for the speedup measured by the openssl speed:

Original:
```
[mraz@leccinum ossl2]$ LD_LIBRARY_PATH=. apps/openssl speed -seconds 1 -evp aes-128-gcm
Doing AES-128-GCM for 1s on 16 size blocks: 40193219 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 64 size blocks: 19933155 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 256 size blocks: 12005343 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 1024 size blocks: 4264277 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 8192 size blocks: 609969 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 16384 size blocks: 308704 AES-128-GCM's in 1.00s
Doing get-info for 1s on 16 size blocks: 11193804 get-info's in 1.00s
version: 3.1.0-dev
built on: Wed Dec  8 17:45:44 2021 UTC
options: bn(64,64)
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DPEDANTIC -pedantic -Wno-long-long -DUNUSEDRESULT_DEBUG -Wall -Wmissing-declarations -Wextra
 -Wno-unused-parameter -Wno-missing-field-initializers -Wswitch -Wsign-compare -Wshadow -Wformat -Wtype-limits -Wundef -Werror -Wmissing-prototypes -Wstric
t-prototypes -g -O2 -Werror -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DZLIB -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ed8320b078bffff:0x400004219c91a9
The 'numbers' are in 1000s of bytes per second processed.
type             16 bytes     64 bytes    256 bytes   1024 bytes   8192 bytes  16384 bytes
AES-128-GCM     643091.50k  1275721.92k  3073367.81k  4366619.65k  4996866.05k  5057806.34k
get-info        179100.86k        0.00         0.00         0.00         0.00         0.00
```

Patched:
```
[mraz@leccinum ossl2]$ LD_LIBRARY_PATH=. apps/openssl speed -seconds 1 -evp aes-128-gcm
Doing AES-128-GCM for 1s on 16 size blocks: 40407762 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 64 size blocks: 19804995 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 256 size blocks: 11938205 AES-128-GCM's in 0.99s
Doing AES-128-GCM for 1s on 1024 size blocks: 4244186 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 8192 size blocks: 604449 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 16384 size blocks: 305132 AES-128-GCM's in 1.00s
Doing get-info for 1s on 16 size blocks: 26123854 get-info's in 1.00s
version: 3.1.0-dev
built on: Wed Dec  8 17:45:44 2021 UTC
options: bn(64,64)
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DPEDANTIC -pedantic -Wno-long-long -DUNUSEDRESULT_DEBUG -Wall -Wmissing-declarations -Wextra
CPUINFO: OPENSSL_ia32cap=0x7ed8320b078bffff:0x400004219c91a9
The 'numbers' are in 1000s of bytes per second processed.
type             16 bytes     64 bytes    256 bytes   1024 bytes   8192 bytes  16384 bytes
AES-128-GCM     646524.19k  1267519.68k  3087050.99k  4346046.46k  4951646.21k  4999282.69k
get-info        417981.66k        0.00         0.00         0.00         0.00         0.00.
```

So the patched code is more than twice as fast.