@t8m 2021-12-16
I like the more generic ptrie approach more unfortunately when I cherry picked the openssl speed patch from this PR and applied it over your ptrie PR @paulidale, this was the result:
```
[mraz@leccinum ossl2]$ LD_LIBRARY_PATH=. apps/openssl speed -seconds 1 -evp aes-128-gcm
Doing AES-128-GCM for 1s on 16 size blocks: 39955098 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 64 size blocks: 20472995 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 256 size blocks: 11662145 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 1024 size blocks: 4235636 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 8192 size blocks: 607921 AES-128-GCM's in 1.00s
Doing AES-128-GCM for 1s on 16384 size blocks: 308065 AES-128-GCM's in 1.00s
Doing get-info for 1s on 16 size blocks: 9850801 get-info's in 1.00s
version: 3.1.0-dev
built on: Thu Dec 16 10:52:52 2021 UTC
options: bn(64,64)
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DPEDANTIC -pedantic -Wno-long-long -DUNUSEDRESULT_DEBUG -Wall -Wmissing-declarations -Wextra -Wno-unused-parameter -Wno-missing-field-initializers -Wswitch -Wsign-compare -Wshadow -Wformat -Wtype-limits -Wundef -Werror -Wmissing-prototypes -Wstrict-prototypes -g -O2 -Werror -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DZLIB -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ed8320b078bffff:0x400004219c91a9
The 'numbers' are in 1000s of bytes per second processed.
type             16 bytes     64 bytes    256 bytes   1024 bytes   8192 bytes  16384 bytes
AES-128-GCM     639281.57k  1310271.68k  2985509.12k  4337291.26k  4980088.83k  5047336.96k
get-info        157612.82k        0.00         0.00         0.00         0.00         0.00 
```