@t8m 2021-12-16
So as you can see calling just EVP_CIPHER_CTX_get_key_length and EVP_CIPHER_CTX_get_iv_length is still slower than original unpatched code.