@paulidale 2021-12-17
We need to make `EVP_CIPHER_CTX_get_key_length` and `EVP_CIPHER_CTX_get_iv_length` calls caching.