@t8m 2021-12-17
> We need to make `EVP_CIPHER_CTX_get_key_length` and `EVP_CIPHER_CTX_get_iv_length` calls caching.

How do you do that? Also that helps only if ctx is reused. I am not sure if that was the case in the original issue that triggered this development.