@paulidale 2021-12-17
There was some reuse I believe.
Or repeated calls to these functions on the same context.