@mattcaswell 2022-01-18
OTC: Instead of this approach we should just cache the results of `EVP_CIPHER_CTX_get_key_length()` etc